agent_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
numpy==2.2.4
scikit-learn==1.6.1
pyyaml==6.0.2
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
